pyproject.toml
[tool.poetry]
name = "user-management"
version = "0.1.0"
description = ""
authors = ["Vinh Bui <vinhbui.tg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
alembic = "^1.9.0"
pydantic = "^1.10.2"
psycopg2-binary = "^2.9.5"
sendgrid = "^6.9.7"
redis = "^4.4.0"
SQLAlchemy = "^1.4.45"
celery = "^5.2.7"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
